+++ src/jsinputdto.ts
@@ -1,16 +1,16 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 
 import { DefaultSettings } from "./settings";
-
-
 interface IJSInputDto
 {
     XMLString: string;
     IndentLength?: number;
     UseSingleQuotes?: boolean;
     UseSelfClosingTags?: boolean;
+    AllowSingleQuoteInAttributeValue?: boolean;
+    AddSpaceBeforeSelfClosingTag?: boolean;
+    WrapCommentTextWithSpaces?: boolean;
 }
-
 
 export class JSInputDTO
 {
@@ -18,16 +18,24 @@ export class JSInputDTO
     IndentLength: number;
     UseSingleQuotes: boolean;
     UseSelfClosingTags: boolean;
+    AllowSingleQuoteInAttributeValue: boolean;
+    AddSpaceBeforeSelfClosingTag: boolean;
+    WrapCommentTextWithSpaces: boolean;
 
-    constructor(XMLString: string,
-        IndentLength: number = DefaultSettings.IndentLength,
-        UseSingleQuotes: boolean = DefaultSettings.UseSingleQuotes,
-        UseSelfClosingTags: boolean = DefaultSettings.UseSelfClosingTags)
+    constructor(xmlString: string,
+        indentLength: number = DefaultSettings.IndentLength,
+        useSingleQuotes: boolean = DefaultSettings.UseSingleQuotes,
+        useSelfClosingTags: boolean = DefaultSettings.UseSelfClosingTags,
+        allowSingleQuoteInAttributeValue: boolean = DefaultSettings.AllowSingleQuoteInAttributeValue,
+        addSpaceBeforeSelfClosingTag: boolean = DefaultSettings.AddSpaceBeforeSelfClosingTag,
+        wrapCommentTextWithSpaces: boolean = DefaultSettings.WrapCommentTextWithSpaces)
     {
-        this.XMLString = XMLString;
-        this.IndentLength = IndentLength;
-        this.UseSingleQuotes = UseSingleQuotes;
-        this.UseSelfClosingTags = UseSelfClosingTags;
+        this.XMLString = xmlString;
+        this.IndentLength = indentLength;
+        this.UseSingleQuotes = useSingleQuotes;
+        this.UseSelfClosingTags = useSelfClosingTags;
+        this.AllowSingleQuoteInAttributeValue = allowSingleQuoteInAttributeValue;
+        this.AddSpaceBeforeSelfClosingTag = addSpaceBeforeSelfClosingTag;
+        this.WrapCommentTextWithSpaces = wrapCommentTextWithSpaces;
     }
-
 }