+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Preview
 
+### 0.9.0 : 07-Jul-2021
+- Added settings : 
+    - `prettyxml.settings.allowSingleQuoteInAttributeValue`
+    - `prettyxml.settings.addSpaceBeforeSelfClosingTag`
+    - `prettyxml.settings.wrapCommentTextWithSpaces`
+
+- Updated **Readme.md**.  
 
 ### 0.8.1